Switched to ES6 module syntax and jsx template added
